+++ src/core/brsTypes/Coercion.ts
