+++ src/core/brsTypes/components/RoCompositor.ts
@@ -14,7 +14,7 @@ import {
     drawObjectToComponent,
     getDimensions,
     releaseCanvas,
-} from "../IfDraw2d";
+} from "../Draw2D";
 
 export class RoCompositor extends BrsComponent implements BrsValue {
     readonly kind = ValueKind.Object;
