+++ src/core/brsTypes/components/RoFunction.ts
@@ -4,7 +4,7 @@ import { Callable, StdlibArgument } from "../Callable";
 import { Interpreter } from "../../interpreter";
 import { BrsType } from "..";
 import { Unboxable } from "../Boxing";
-import { ifToStr } from "../interfaces/ifToStr";
+import { ifToStr } from "../interfaces/ifToStrModule";
 
 export class RoFunction extends BrsComponent implements BrsValue, Unboxable {
     readonly kind = ValueKind.Object;
